Rename the last test-named module that collects no tests
Closes #18.

`examples/data/test_data.py` matched pytest's other default pattern, `test_*.py`.
[JOSS S8] swept the `*_test.py` suffix form and this one survived, so the finding
was still live in the module that downloads the price fixture.

It collected nothing but was still *imported* at collection — the mechanism
behind the two CI failures a module-level `yfinance` import has caused here.
Confirmed rather than assumed, before and after: appending a `raise` and running
`pytest --collect-only` errored on the old name and does not on the new one.

Renamed to `examples/data/etf_prices_local.py`, following the convention
[JOSS S8] established. Both halves of the module need the author's machine —
`update_test_prices` downloads from Yahoo, `load_test_data` reads `RESOURCE_PATH`
— so `_local` is what it is, even though it doubles as the shared loader the six
`*_local.py` diagnostics import their panel from. The module docstring now says
so, and says why the name matters.

`load_test_data` and `update_test_prices` are unchanged and were never exported
from the package, so only the six import paths move.

The issue's "done when" is now checkable in one line, and checked: 27 files match
either default pattern and all 27 collect at least one test. AGENTS.md states
that invariant rather than only the file count, since the count is what drifted.

Verified on 3.10 / pandas 2.3.3 and 3.11 / pandas 3.0.5: 632 tests pass on both,
ruff stack invariants green, interrogate at 100%.
