+++ optimalportfolios/optimization/tests/max_diversification_local.py
@@ -83,7 +83,7 @@ def run_local_test(local_test: LocalTests):
 
     elif local_test == LocalTests.MAX_DIVERSIFICATION_ROLLING:
         import qis as qis
-        from optimalportfolios.examples.data.test_data import load_test_data
+        from optimalportfolios.examples.data.etf_prices_local import load_test_data
         prices = load_test_data()
         prices = prices.loc['2000':, :]
